Made a small change on your menu
In your menu, if the user input is something that is not a number between 1 and 8, the menu is sent into an infinite loop, continually printing the menu. To fix this, I added a small while loop inside of your while loop, which runs whenever the input is not a number between 1 and 8. This will just tell the user to enter a correct input.